+++ R.swift/Util/Struct+ChildValidation.swift
@@ -1,0 +1,48 @@
+//
+//  Struct+ChildValidation.swift
+//  R.swift
+//
+//  Created by Mathijs Kadijk on 22-12-15.
+//  Copyright © 2015 Mathijs Kadijk. All rights reserved.
+//
+
+import Foundation
+
+extension Struct {
+  /// Implements the Validatable protocol on this and child struct if one or more childs already implement the 
+  /// Validatable protocol. The newly created validate methods will call their child validate methods.
+  func addChildStructValidationMethods() -> Struct {
+    if implements.contains(Type.Validatable) {
+      return self
+    }
+
+    let childStructs = structs
+      .map { $0.addChildStructValidationMethods() }
+
+    let validatableStructs = childStructs
+      .filter { $0.implements.contains(Type.Validatable) }
+
+    guard validatableStructs.count > 0 else {
+      return self
+    }
+
+    var outputStruct = self
+    outputStruct.structs = childStructs
+    outputStruct.implements.append(Type.Validatable)
+    outputStruct.functions.append(
+      Function(
+        isStatic: true,
+        name: "validate",
+        generics: nil,
+        parameters: [],
+        doesThrow: true,
+        returnType: Type._Void,
+        body: validatableStructs
+          .map { "try \($0.type).validate()" }
+          .joinWithSeparator("\n")
+      )
+    )
+    
+    return outputStruct
+  }
+}
